requirements.txt
google-cloud-datastore==1.12.0
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-cloud-bigquery==1.25.0
google-cloud-pubsub==1.6.0
pandas==1.1.0
pandas-gbq==0.13.2
pyarrow==1.0.0
yarl==1.4.2 # Workaround because Slack fails to install in Cloud Function
slackclient==2.7.2
pyyaml==5.3.1
pytest==5.4.1
pycodestyle==2.6.0
pytest-cov==2.8.1